agent/pyproject.toml
[project]
name = "phoenix-agent"
version = "5.0.0"
requires-python = ">=3.12"
license = "AGPL-3.0-only"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pygithub>=2.5.0",
    "openhands-ai>=0.38.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "anthropic>=0.40.0",
    "aiosqlite>=0.20.0",
]

[project.scripts]
phoenix-agent    = "server:serve"
phoenix-semantic = "semantic_server:serve_semantic"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
# E/F = pycodestyle + pyflakes, I = isort, UP = pyupgrade, B = bugbear
select = ["E", "F", "I", "UP", "B"]
ignore = [
    "E501",   # line too long — handled by the formatter
    "B008",   # function calls in default args (FastAPI depends-on pattern)
]

[tool.ruff.lint.isort]
known-first-party = ["agent", "app", "config", "db", "models", "registry", "routes"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
